Update UserDTOs and UserController in Identity microservice and update Assignment/MaterialController in CourseContent microservice

+++ EducationPlatform.CourseContent/CourseContent.Core/Services/AssignmentService.cs
@@ -72,6 +72,11 @@ namespace CourseContent.Core.Services
 
         public async Task<Result<string, Error>> RemoveRangeAsync(List<int> entities)
         {
+            if (entities.Count == 0)
+            {
+                return Result.Failure<string, Error>(Errors.General.NotRecords());
+            }
+
             await _unitOfWork.AssignmentRepository.RemoveRange(entities);
             await _unitOfWork.CompleteAsync();
             return Result.Success<string, Error>("Deleted was successful");
